+++ src/Control/Exception/Annotated/UnliftIO.hs
@@ -1,14 +1,14 @@
-{-# language ExplicitForAll #-}
+{-# LANGUAGE ExplicitForAll #-}
 
 -- | This module presents the same interface as
 -- "Control.Exception.Annotated", but uses 'MonadUnliftIO' instead of
--- 'MonadCatch' or 'MonadThrow'.
+-- 'Control.Monad.Catch.MonadCatch' or 'Control.Monad.Catch.MonadThrow'.
 --
 -- @since 0.1.2.0
 module Control.Exception.Annotated.UnliftIO
     ( -- * The Main Type
       AnnotatedException(..)
-    , new
+    , exceptionWithCallStack
     , throwWithCallStack
     -- * Annotating Exceptions
     , checkpoint
@@ -56,38 +56,40 @@ import Control.Monad.IO.Unlift
 import GHC.Stack
 
 -- | Like 'Catch.throwWithCallStack', but uses 'MonadIO' instead of
--- 'MonadThrow'.
+-- 'Control.Monad.Catch.MonadThrow'.
 --
 -- @since 0.1.2.0
 throwWithCallStack
     :: forall e m a. (MonadIO m, Exception e, HasCallStack)
     => e -> m a
-throwWithCallStack = liftIO . Catch.throwWithCallStack
+throwWithCallStack = liftIO . withFrozenCallStack Catch.throwWithCallStack
 
--- | Like 'Catch.throw', but uses 'MonadIO' instead of 'MonadThrow'.
+-- | Like 'Catch.throw', but uses 'MonadIO' instead of 'Control.Monad.Catch.MonadThrow'.
 --
 -- @since 0.1.2.0
-throw :: forall e m a. (MonadIO m, Exception e) => e -> m a
-throw = liftIO . Catch.throw
+throw :: forall e m a. (MonadIO m, Exception e, HasCallStack) => e -> m a
+throw = liftIO . withFrozenCallStack Catch.throw
 
--- | Like 'Catch.checkpoint', but uses 'MonadUnliftIO' instead of 'MonadCatch'.
+-- | Like 'Catch.checkpoint', but uses 'MonadUnliftIO' instead of 'Control.Monad.Catch.MonadCatch'.
 --
 -- @since 0.1.2.0
-checkpoint :: forall m a. (MonadUnliftIO m) => Annotation -> m a -> m a
+checkpoint :: forall m a. (MonadUnliftIO m, HasCallStack) => Annotation -> m a -> m a
 checkpoint ann action = withRunInIO $ \runInIO ->
-    liftIO $ Catch.checkpoint ann (runInIO action)
+    liftIO $ withFrozenCallStack (Catch.checkpoint ann) (runInIO action)
 
 -- | Like 'Catch.checkpointMany', but uses 'MonadUnliftIO' instead of
--- 'MonadCatch'.
+-- 'Control.Monad.Catch.MonadCatch'.
 --
 -- @since 0.1.2.0
-checkpointMany :: forall m a. (MonadUnliftIO m) => [Annotation] -> m a -> m a
+checkpointMany :: forall m a. (MonadUnliftIO m, HasCallStack) => [Annotation] -> m a -> m a
 checkpointMany anns action =
     withRunInIO $ \runInIO ->
-        liftIO $ Catch.checkpointMany anns (runInIO action)
+        liftIO $ withFrozenCallStack Catch.checkpointMany anns (runInIO action)
 
 -- | Like 'Catch.checkpointCallStackWith', but uses 'MonadUnliftIO' instead of
--- 'MonadCatch'.
+-- 'Control.Monad.Catch.MonadCatch'.
+--
+-- Deprecated in 0.2.0.0 as it is now an alias for 'checkpointMany'.
 --
 -- @since 0.1.2.0
 checkpointCallStackWith
@@ -95,9 +97,11 @@ checkpointCallStackWith
     => [Annotation] -> m a -> m a
 checkpointCallStackWith anns action =
     withRunInIO $ \runInIO ->
-        liftIO $ Catch.checkpointCallStackWith anns (runInIO action)
+        liftIO $ withFrozenCallStack Catch.checkpointCallStackWith anns (runInIO action)
 
--- | Like 'Catch.catch', but uses 'MonadUnliftIO' instead of 'MonadCatch'.
+{-# DEPRECATED checkpointCallStackWith "As of annotated-exception-0.2.0.0, this is an alias for checkpointMany" #-}
+
+-- | Like 'Catch.catch', but uses 'MonadUnliftIO' instead of 'Control.Monad.Catch.MonadCatch'.
 --
 -- @since 0.1.2.0
 catch
@@ -109,7 +113,7 @@ catch action handler =
     withRunInIO $ \runInIO ->
         liftIO $ Catch.catch (runInIO action) (\e -> runInIO $ handler e)
 
--- | Like 'Catch.tryAnnotated' but uses 'MonadUnliftIO' instead of 'MonadCatch'.
+-- | Like 'Catch.tryAnnotated' but uses 'MonadUnliftIO' instead of 'Control.Monad.Catch.MonadCatch'.
 --
 -- @since 0.1.2.0
 tryAnnotated
@@ -120,7 +124,7 @@ tryAnnotated action =
     withRunInIO $ \runInIO ->
         liftIO $ Catch.tryAnnotated (runInIO action)
 
--- | Like 'Catch.try' but uses 'MonadUnliftIO' instead of 'MonadCatch'.
+-- | Like 'Catch.try' but uses 'MonadUnliftIO' instead of 'Control.Monad.Catch.MonadCatch'.
 --
 -- @since 0.1.2.0
 try
@@ -131,7 +135,7 @@ try action =
     withRunInIO $ \runInIO ->
         liftIO $ Catch.try (runInIO action)
 
--- | Like 'Catch.catches', bt uses 'MonadUnliftIO' instead of 'MonadCatch'.
+-- | Like 'Catch.catches', bt uses 'MonadUnliftIO' instead of 'Control.Monad.Catch.MonadCatch'.
 --
 -- @since 0.1.2.0
 catches
@@ -143,4 +147,3 @@ catches action handlers =
     withRunInIO $ \runInIO -> do
         let f (Handler k) = Handler (\e -> runInIO (k e))
         liftIO $ catches (runInIO action) (map f handlers)
-  where