+++ src/Data/Annotation.hs
@@ -7,17 +7,53 @@
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
 
--- | An 'Annotation' is attached to a 'LocatedException'. They're
+-- | An 'Annotation' is attached to an 'Control.Exception.Annotated.AnnotatedException'. They're
 -- essentially a dynamically typed value with a convenient 'IsString'
--- instance. I'd recommend using something like @Data.Aeson.Value@ or
--- possibly something more strongly typed.
+-- instance.
+--
+-- When integrating into your own application, you will likely want to do more
+-- than just have the 'String' you get from 'show'ing the 'Annotation'. You can
+-- do this by creating a special wrapper type that carries a class constraint.
+-- This allows you to also pluck out the 'Annotation's from your library or
+-- executable independently and treat them differently from unknonwn
+-- annotations.
+--
+-- As an example, here's one that requires a 'Data.Aeson.ToJSON' constraint on the
+-- underlying value. This means that you can convert any annotated value to
+-- JSON, and then use that JSON in bug reports or logging.
+--
+-- @
+-- data JsonAnnotation where
+--   JsonAnnotation :: (ToJSON a, Typeable a) => a -> JsonAnnotation
+--
+-- instance Show JsonANnotation where
+--   show (JsonAnnotation a) = concat
+--      [ "(JsonAnnotation ("
+--      , show (toJSON a)
+--      , "))"
+--      ]
+--
+-- jsonCheckpoint :: (Typeable a, ToJSON a, 'HasCallStack', MonadCatch m) => a -> m a -> m a
+-- jsonCheckpoint val = 'withFrozenCallStack' checkpoint (JsonAnnotation val)
+-- @
+--
+-- When handling the @['Annotation']@ carried on the
+-- 'Control.Exception.Annotated.AnnotatedException', you can use
+-- 'tryAnnotations' to pick out the JSON annotations.
+--
+-- @
+-- jsonAnnotations :: [Annotation] -> ([JsonAnnotation], [Annotation])
+-- jsonAnnotations = tryAnnotations
+-- @
+--
+-- and handle them however you please.
+--
+-- @since 0.1.0.0
 module Data.Annotation
     ( module Data.Annotation
     , module Data.Proxy
     ) where
 
-import GHC.Stack
-import Data.Dynamic
 import Data.Either
 import Data.Maybe
 import Data.Proxy
@@ -26,6 +62,7 @@ import qualified Data.Set as Set
 import Data.String
 import qualified Data.Text as Text
 import Data.Typeable
+import GHC.Stack
 
 -- | The constraints that the value inside an 'Annotation' must have.
 --
@@ -33,11 +70,12 @@ import Data.Typeable
 -- information out of it.
 --
 -- @since 0.1.0.0
-type AnnC a = (Typeable a, Eq a, Show a)
+type AnnC a = (Typeable a, Show a)
 
 -- | An 'Annotation' is a wrapper around a value that includes a 'Typeable'
--- constraint so we can later unpack it. It is essentially a 'Dynamic, but
--- we also include 'Show' and 'Eq' so it's more useful.
+-- constraint so we can later unpack it. It is essentially a 'Data.Dynamic.Dynamic', but
+-- we also include 'Show' so that you can always fall back to simply 'show'ing
+-- the 'Annotation' if it is otherwise unrecognized.
 --
 -- @since 0.1.0.0
 data Annotation where
@@ -49,19 +87,13 @@ data Annotation where
 -- |
 --
 -- @since 0.1.0.0
-instance Eq Annotation where
-    Annotation (a :: a) == Annotation (b :: b) =
-        case eqT @a @b of
-            Just Refl ->
-                a == b
-            Nothing ->
-                False
-
--- |
---
--- @since 0.1.0.0
 instance Show Annotation where
-    show (Annotation a) = show a
+    showsPrec p (Annotation a) =
+        showParen (p > 10) $
+            showString "Annotation @"
+            . showsPrec 11 (typeOf a)
+            . showString " "
+            . showsPrec 11 a
 
 -- |
 --
@@ -145,32 +177,41 @@ mapMaybeAnnotation f ann =
 -- | A wrapper type for putting a 'CallStack' into an 'Annotation'. We need
 -- this because 'CallStack' does not have an 'Eq' instance.
 --
+-- Deprecated in 0.2.0.0 since you can just put a 'CallStack' directly in an
+-- 'Annotation' now that we have no need for an 'Eq' constraint on it.
+--
 -- @since 0.1.0.0
 newtype CallStackAnnotation = CallStackAnnotation
     { unCallStackAnnotation :: [(String, SrcLoc)]
     }
     deriving (Eq, Show)
 
+{-# DEPRECATED CallStackAnnotation "You can just use `CallStack` directly now." #-}
+
 -- | Grab an 'Annotation' corresponding to the 'CallStack' that is
 -- currently in scope.
 --
 -- @since 0.1.0.0
 callStackAnnotation :: HasCallStack => Annotation
-callStackAnnotation = callStackToAnnotation callStack
+callStackAnnotation = Annotation callStack
 
 -- | Stuff a 'CallStack' into an 'Annotation' via the 'CallStackAnnotation'
 -- newtype wrapper.
 --
 -- @since 0.1.0.0
 callStackToAnnotation :: CallStack -> Annotation
-callStackToAnnotation cs = Annotation $ CallStackAnnotation $ getCallStack cs
+callStackToAnnotation = Annotation
 
--- | Attempt to convert an 'Annotation' back into a 'CallStack'.
+-- | Convert the legacy 'CallStackAnnotation' into a 'CallStack'.
+--
+-- Deprecated in 0.2.0.0 since you can use 'CallStack' directly.
 --
 -- @since 0.1.0.0
 callStackFromAnnotation :: CallStackAnnotation -> CallStack
 callStackFromAnnotation ann =
     fromCallSiteList $ unCallStackAnnotation ann
+
+{-# DEPRECATED callStackFromAnnotation "You can use 'CallStack' directly in annotations as of 0.2.0.0." #-}
 
 -- | Extract the 'CallStack's from the @['Annotation']@. Any 'Annotation'
 -- not corresponding to a 'CallStack' will be in the second element of the
@@ -178,8 +219,7 @@ callStackFromAnnotation ann =
 --
 -- @since 0.1.0.0
 callStackInAnnotations :: [Annotation] -> ([CallStack], [Annotation])
-callStackInAnnotations anns =
-    let (callStacks, rest) =
-            tryAnnotations anns
-    in
-        (fmap callStackFromAnnotation callStacks, rest)
+callStackInAnnotations =
+    tryAnnotations
+
+{-# DEPRECATED callStackInAnnotations "You can just use 'tryAnnotations' directly as of 0.2.0.0." #-}