+++ test/Data/AnnotationSpec.hs
@@ -1,23 +1,35 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
 
 module Data.AnnotationSpec where
 
 import Data.Annotation
+import Data.Typeable
 import Test.Hspec
 
 spec :: Spec
 spec = do
-    describe "Eq" $ do
-        it "works for equal values" $ do
-            toAnnotation "hello" == toAnnotation "hello"
-        it "works for non-equal values of same type" $ do
-            toAnnotation "a" /= toAnnotation "b"
-        it "works for values of different types" $ do
-            toAnnotation (1 :: Int) /= toAnnotation "a"
     describe "Show" $ do
-        it "works" $ do
-            show (toAnnotation (3 :: Int))
+        it "includes type information" $ do
+            show (Annotation @Int 3)
                 `shouldBe`
-                    "3"
+                    "Annotation @Int 3"
+        it "uses parens" $ do
+            show (Annotation @(Maybe Int) (Just 3))
+                `shouldBe`
+                    "Annotation @(Maybe Int) (Just 3)"
+        it "is cool with strings" $ do
+            show (Annotation @String "Hello")
+                `shouldBe`
+                    "Annotation @[Char] \"Hello\""
+
+instance Eq Annotation where
+    Annotation (a :: a) == Annotation (b :: b) =
+        case eqT @a @b of
+            Just Refl ->
+                show a == show b
+            Nothing ->
+                False